+++ pool/app/assignment/repo/repo.ml
@@ -151,7 +151,10 @@ module Sql = struct
     let joins =
       {sql|
         LEFT JOIN pool_sessions
-        ON pool_assignments.session_uuid = pool_sessions.uuid
+        ON
+          pool_assignments.session_uuid = pool_sessions.uuid
+        AND
+          pool_sessions.canceled_at IS NULL
       |sql}
     in
     {sql|
@@ -252,6 +255,8 @@ module Sql = struct
         $9,
         $10
       )
+      ON DUPLICATE KEY UPDATE
+        marked_as_deleted = 0
     |sql}
     |> RepoEntity.t ->. Caqti_type.unit
   ;;