+++ pool/app/message_template/repo/repo_sql.ml
@@ -244,6 +244,27 @@ let find_all_of_entity_by_label pool entity_uuid label =
     (Pool_common.Id.value entity_uuid, Label.show label)
 ;;
 
+let find_by_label_and_language_request =
+  let open Caqti_request.Infix in
+  Format.asprintf
+    {sql|
+    %s
+    WHERE
+      pool_message_templates.label = ?
+    AND
+      pool_message_templates.language = ?
+    |sql}
+    select_sql
+  |> Caqti_type.(tup2 string string) ->! RepoEntity.t
+;;
+
+let find_by_label_and_language pool language label =
+  Utils.Database.find_opt
+    (Pool_database.Label.value pool)
+    find_by_label_and_language_request
+    (Entity.Label.show label, Pool_common.Language.show language)
+;;
+
 let find_request =
   let open Caqti_request.Infix in
   Format.asprintf
@@ -266,8 +287,7 @@ let find pool id =
 
 let insert_default_if_not_exists pool t =
   let open Utils.Lwt_result.Infix in
-  find_by_label_to_send pool t.Entity.language t.Entity.label
-  ||> CCResult.to_opt
+  find_by_label_and_language pool t.Entity.language t.Entity.label
   >|> function
   | None -> insert pool t
   | Some _ -> Lwt.return ()