+++ pool/cqrs_command/assignment_command.ml
@@ -16,7 +16,6 @@ module Create : sig
   type t =
     { contact : Contact.t
     ; sessions : Session.Public.t list
-    ; waiting_list : Waiting_list.t option
     ; experiment : Experiment.Public.t
     }
 
@@ -30,7 +29,6 @@ end = struct
   type t =
     { contact : Contact.t
     ; sessions : Session.Public.t list
-    ; waiting_list : Waiting_list.t option
     ; experiment : Experiment.Public.t
     }
 
@@ -67,20 +65,13 @@ end = struct
              in
              Assignment.Created create |> Pool_event.assignment)
       in
-      let delete_events =
-        match command.waiting_list with
-        | None -> []
-        | Some waiting_list ->
-          [ Waiting_list.Deleted waiting_list |> Pool_event.waiting_list ]
-      in
       let increase_num_events =
         Contact.NumAssignmentsIncreasedBy
           (command.contact, CCList.length command.sessions)
         |> Pool_event.contact
       in
       Ok
-        (delete_events
-         @ create_events
+        (create_events
          @ [ increase_num_events ]
          @ [ Email.Sent confirmation_email |> Pool_event.email ])
   ;;
@@ -252,8 +243,6 @@ end = struct
          @ [ Contact.NumAssignmentsIncreasedBy
                (contact, CCList.length command.sessions)
              |> Pool_event.contact
-           ; Waiting_list.Deleted command.waiting_list
-             |> Pool_event.waiting_list
            ; Email.Sent confirmation_email |> Pool_event.email
            ])
   ;;